fix warning for %u formatting
svn path=/trunk/; revision=20767

+++ libraries/psx/psx_malloc.c
@@ -183,8 +183,8 @@ F77_SUBROUTINE(psx_malloc)( INTEGER(size), POINTER(pntr), INTEGER(status) )
       *pntr = (F77_POINTER_TYPE)0;
       *status = PSX__NOALL;
       sprintf( errbuf, 
-         "Failed to allocate space with malloc. %u bytes requested",
-         (size_t)(*size) );
+         "Failed to allocate space with malloc. %lu bytes requested",
+	       (unsigned long)(*size) );
       psx1_rep_c( "PSX_MALLOC_NOALL", errbuf, status );
    }
